Move pallets middlewares for running tests to their own directory
