+++ bin/test/tasks/exit.rb
@@ -3,16 +3,17 @@
 class Test::Tasks::Exit < Pallets::Task
   include Test::TaskHelpers
 
+  # rubocop:disable Metrics/MethodLength
   def run
     overall_finish_time = Process.clock_gettime(Process::CLOCK_MONOTONIC)
 
     # print the run times and exit codes for all jobs in this pallets run
     puts("\n")
     max_job_name_length =
-      TaskResultTrackingMiddleware.job_results.keys.
+      Test::Middleware::TaskResultTrackingMiddleware.job_results.keys.
         map { _1.sub('Test::Tasks::', '') }.map(&:size).
         max
-    TaskResultTrackingMiddleware.job_results.
+    Test::Middleware::TaskResultTrackingMiddleware.job_results.
       sort_by { |_task_name, result_hash| [result_hash[:exit_code], result_hash[:run_time]] }.
       each do |task_name, result_hash|
         exit_code = result_hash[:exit_code]
@@ -32,7 +33,8 @@ class Test::Tasks::Exit < Pallets::Task
     overall_start_time = Test::Runner.start_time
     overall_wall_clock_time = overall_finish_time - overall_start_time
     puts("Wall clock elapsed time: #{overall_wall_clock_time.round(3).to_s.yellow}")
-    total_task_time = TaskResultTrackingMiddleware.job_results.map { _2[:run_time] }.inject(:+)
+    total_task_time =
+      Test::Middleware::TaskResultTrackingMiddleware.job_results.map { _2[:run_time] }.inject(:+)
     puts("Total task time: #{total_task_time.round(3).to_s.yellow}")
     puts("Parallism: #{total_task_time.fdiv(overall_wall_clock_time).round(3).to_s.yellow}")
 
@@ -40,4 +42,5 @@ class Test::Tasks::Exit < Pallets::Task
     puts("Exiting with code #{Test::Runner.exit_code}")
     exit(Test::Runner.exit_code)
   end
+  # rubocop:enable Metrics/MethodLength
 end
